Functions Get Duplicate numbers in a specific row and column added
For solving the problem with the Hill-Climbing algorithm, Our approach is to traverse the diagonal of matrix to only check and solve the problem once.

That's the reason which we defined the function getRow_Column like this. we use only an input and get the specific Row and Column of that index in the Matrix.

Then we defined the func findDuplicate to get the number whhich are duplicate in that row and column.
===========================
Our next step is to define our heuristic func (h(n)). At this time We will define the heuristic as the number of duplicate numbers in the whole matrix.